separação da responsabilidade das rotas e prévia do método para criação da tela de cadastro
