+++ package.json
@@ -10,13 +10,18 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.0.5",
+    "@react-navigation/drawer": "^6.1.4",
     "@react-navigation/native": "^6.0.2",
     "@react-navigation/native-stack": "^6.1.0",
     "react": "17.0.2",
     "react-native": "0.65.1",
-    "react-native-safe-area-context": "^3.3.1",
-    "react-native-screens": "^3.6.0",
-    "react-native-vector-icons": "^8.1.0"
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-screens": "^3.7.0",
+    "react-native-vector-icons": "^8.1.0",
+    "react-redux": "5.0.7",
+    "redux": "4.0.0",
+    "redux-thunk": "2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.15.0",
