+++ src/router.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import Login from './screens/Login';
-import Registry from './screens/Registry';
+import Register from './screens/Register';
 
 const {Navigator, Screen} = createNativeStackNavigator();
 
@@ -17,7 +17,7 @@ const Router = () => {
           },
         }}>
         <Screen name="Login" component={Login} />
-        <Screen name="Register" component={Registry} />
+        <Screen name="Register" component={Register} />
       </Navigator>
     </NavigationContainer>
   );
