+++ src/screens/Login.js
@@ -1,15 +1,16 @@
 /* eslint-disable react-native/no-inline-styles */
-import React, {useState} from 'react';
-import {View, Text, StyleSheet} from 'react-native';
+import React, { useState } from 'react';
+import { View, Text, StyleSheet } from 'react-native';
 import Logo from '../components/Logo';
 import Input from '../components/Input';
 import ButtonPrimary from '../components/ButtonPrimary';
-import {theme} from '../global/styles/theme';
+import { theme } from '../global/styles/theme';
 import ButtonSecondary from '../components/ButtonSecondary';
 import ButtonGoogleLogin from '../components/ButtonGoogleLogin';
 import ButtonFacebookLogin from '../components/ButtonFacebookLogin';
 
-export function Login() {
+
+export function Login({navigation}) {
   const [email, onChangeEmail] = useState();
   const [password, onChangePassword] = useState();
 
@@ -21,21 +22,23 @@ export function Login() {
           placeholder={'Email'}
           onChangeText={onChangeEmail}
           //react-native/no-inline-styles
-          style={{marginTop: 40}}
+          style={{ marginTop: 40 }}
         />
         <Input
           placeholder={'Senha'}
           onChangeText={onChangePassword}
-          style={{marginTop: 8}}
+          style={{ marginTop: 8 }}
         />
         <Text>{email}</Text>
         <Text>{password}</Text>
         <ButtonPrimary>ENTRAR</ButtonPrimary>
-        <ButtonSecondary>CADASTRAR</ButtonSecondary>
-        <View style={{flexDirection: 'row', alignItems: 'center'}}>
+        <ButtonSecondary onPress={() => {
+        navigation.navigate('Register')
+        }}>CADASTRAR</ButtonSecondary>
+        <View style={{ flexDirection: 'row', alignItems: 'center' }}>
           <View style={styles.line} />
           <View>
-            <Text style={{width: 25, textAlign: 'center', fontSize: 8}}>
+            <Text style={{ width: 25, textAlign: 'center', fontSize: 8 }}>
               ou
             </Text>
           </View>
